+++ src/components/layouts/MainLayout/MainLayout.tsx
@@ -4,7 +4,7 @@ import { Sidebar } from "@/components/layouts/Sidebar";
 import { Props } from "./Layout.types";
 import Head from "next/head";
 
-const Layout: React.FC<Props> = ({ projectName, children }) => {
+const MainLayout: React.FC<Props> = ({ projectName, children }) => {
   return (
     <div className="min-h-screen bg-gray-100">
       <Head>
@@ -21,4 +21,4 @@ const Layout: React.FC<Props> = ({ projectName, children }) => {
   );
 };
 
-export { Layout };
+export { MainLayout };
