+++ src/components/modules/GenerateCodes/GenerateCodes.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import useExecStatus from "@/hooks/useExecStatus";
 import { Card } from "@/components/layouts/Card";
 import { ExecStatusDisplay } from "@/components/elements/ExecStatusDisplay";
-import useGenerateCodesApi from "@/hooks/useGenerateCodesApi";
+import useCreateSecretCodesApi from "@/hooks/useCreateSecretCodesApi";
 import { parseEther } from "ethers/lib/utils";
 
 const GenerateCodes: React.FC = () => {
@@ -10,7 +10,7 @@ const GenerateCodes: React.FC = () => {
   const [amount, setAmount] = useState<string>("");
   const [execStatus, updateExecStatus, clearExecStatus] = useExecStatus();
 
-  const { sendRequest } = useGenerateCodesApi(updateExecStatus);
+  const { sendRequest } = useCreateSecretCodesApi(updateExecStatus);
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
