+++ src/pages/api/get-merkle-proof.ts
@@ -1,7 +1,7 @@
 import CodesTreeModel from "@/models/CodesTreeModel";
 import { ICodesTree } from "@/models/CodesTreeModel.types";
 import { loadMerkleTree } from "@/utils/merkleTree";
-import connectToDatabase from "@/utils/mongoose";
+import connectToDatabase from "@/services/mongoose";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 const getMerkleProof = async (req: NextApiRequest, res: NextApiResponse) => {
